requirements.txt
flask
google-generativeai
python-dotenv
transformers
torch
numpy
pandas
requests
# For financial and Hinglish models
sentencepiece